% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.r
\name{change}
\alias{change}
\title{changes}
\usage{
change(df_original, df_nouveau, id)
}
\arguments{
\item{df_original}{A dataframe containing the original data. Default is null}

\item{df_nouveau}{A dataframe containing the data to compare with}
}
\description{
This function compare deux dataframes à partir du package daff et trie les comparaisons
retourne une liste de 4 dataframes,
updated: un df avec toutes les lignes mises à jour
added: un df avec toutes les lignes ajoutés
deleted: un df avec toutes les lignes supprimées
}
\examples{
changes()
dep = stringr, dplyr, daff
}
