% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_store.r
\name{col_store}
\alias{col_store}
\title{col_store}
\usage{
col_store(conn, bdd, schema, table_name, drop = FALSE)
}
\arguments{
\item{conn}{A DBI connection}

\item{bdd}{The sql database name}

\item{schema}{The sql schema name}

\item{drop}{default is FALSE, if True, drop the existing index first to create the new one.}

\item{table}{The sql Table name}
}
\description{
This function create a columns stored index on a specified table
}
\examples{
col_store(conn = conn, BDD = 'STATPRODTEMP',Schema = 'GEN',table_name = 'MTCARS',drop = FALSE)
col_store()

}
