% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dates.R
\name{convert.dates}
\alias{convert.dates}
\title{convert.dates}
\usage{
convert.dates(df)
}
\arguments{
\item{df}{un dataframe pour lequel des transformation sont requises}
}
\value{
Retourne un dataframe avec les dates dans le bon format
}
\description{
Cette fonction identifie les vecteur de date de différent format character et les transforme en date
}
\examples{
df <- convert.dates(df)
convert.dates()

}
