% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copie.R
\name{copie}
\alias{copie}
\title{copie}
\usage{
copie(df)
}
\arguments{
\item{df}{Un dataframe}
}
\description{
Cette fonction copie dans le clipboard un dataframe
}
\examples{
copie(df = iris)
copie(df = iris)

}
