% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_table.R
\name{drop_table}
\alias{drop_table}
\title{drop_table}
\usage{
drop_table(conn, bdd, schema, table_name)
}
\arguments{
\item{conn}{A DBI connection}

\item{bdd}{The sql database name}

\item{schema}{The sql schema name}

\item{table}{The sql Table name}
}
\description{
Cette fonction supprime une table spécifiée.
}
\examples{
drop_table(conn = conn, BDD = 'STATPRODTEMP',Schema = 'GEN',table_name = 'MTCARS')
col_store()

}
