% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.r
\name{fact}
\alias{fact}
\title{fact}
\usage{
fact(df, n.distinct = 10)
}
\arguments{
\item{df}{Un dataframe}

\item{n.distinct}{Dans le cas où les facteurs sont en classe caractère, n.distinct permet de définir le nombre
de valeurs distincte maximale qu'une variable caractère peut prendre en étant considérée comme une variable de type facteur.}
}
\description{
Cette fonction affiche toutes les valeurs et le nombre d'occurence des facteurs d'un dataframe
}
\examples{
fact(df = iris, n.distinct = 10)
fact(df = iris, n.distinct = 5)

}
