% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggformat.r
\name{ggformat}
\alias{ggformat}
\title{resumedf}
\usage{
ggformat(
  g,
  title = NULL,
  subtitle = NULL,
  x.label = NULL,
  y.label = NULL,
  caption = NULL,
  tag = NULL,
  xlab.angle = 0,
  ylab.angle = 0
)
}
\arguments{
\item{g}{Un objet ggplot}

\item{title}{Le titre du graphique à passer comme argument à la fonction ggplot2::labs()}

\item{subtitle}{Le sous-titre du graphique à passer comme argument à la fonction ggplot2::labs()}

\item{x.label}{Le titre de l'axe x}

\item{y.label}{Le titre de l'axe y}

\item{caption}{...}

\item{tag}{...}

\item{xlab.angle}{Angle en degré pour une rotation des graduation de l'axe x}

\item{ylab.angle}{Angle en degré pour une rotation des graduation de l'axe y}
}
\value{
Retourne un objet ggplot mis en forme
}
\description{
This function mets a jour les données dans une table si l'id existe dans la table sql server ou alors ajoute les données a partir d'un dataframe
}
\examples{
ggformat(g)

g <- mtcars \%>\%
ggplot(aes(x  = carb, y = ws))+
geom_bar(stat = 'identity')

ggformat(g,
   title = 'Titre de mon graphique',
   subtitle = 'sous - titre',
   caption = 'fig. n',
   x.label = 'carburant',
   y.label = 'weigth')



}
