% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sql.R
\name{import_sql}
\alias{import_sql}
\title{import_sql}
\usage{
import_sql(conn, Schema, TableSQL)
}
\arguments{
\item{conn}{Une connexion DBI}

\item{Schema}{Le nom du schéma sur lequel la table se trouve}

\item{TableSQL}{Le nom de la table SQL}
}
\value{
charge dans l'environnement R la table SQL
}
\description{
Cette fonctio charge dans l'environnement R la une table sql dont le nom et le schéma sont passé en argument.
}
\examples{
import_sql(conn = conn, Schema = 'GEN',TableSQL = '[11001041_EGID]')
import_sql()

}
