% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.r
\name{logs}
\alias{logs}
\title{logs}
\usage{
logs(methode, bdd, schema, table_name)
}
\arguments{
\item{methode}{la méthode de mise à jour: 'upsert','table_create','drop_replace'}

\item{bdd}{The sql database name}

\item{schema}{The sql schema}

\item{table_name}{The sql Table}
}
\description{
Cette fonction ajoute une ligne dans la table de logs des modifications apportées aux tables
}
\examples{
logs(methode, bdd, schema, table_name, user)
logs()
}
