% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modif_df.R
\name{modif_df}
\alias{modif_df}
\title{modif_df}
\usage{
modif_df(df, file.type = "excel")
}
\arguments{
\item{df}{Un dataframe}
}
\description{
Cette fonction ouvre un df dans excel pour apporter des modifications et retourner le df modifié
}
\examples{
df <- modif_df(df = iris)
df <- modif_df(df = iris)

}
