% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_to_int.r
\name{num_to_int}
\alias{num_to_int}
\title{A numeric to integer}
\usage{
num_to_int(df = NULL)
}
\arguments{
\item{df}{A dataframe containing numeric variables that as to be stored as integer. Default is null}
}
\description{
This function transform data type that are trully integer but stored in numeric format into integer format.
}
\examples{
num_to_int()
}
\keyword{cats}
