% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palcol.r
\name{palcol}
\alias{palcol}
\title{palcol}
\usage{
palcol(n = 12, show = F, type = "cat20")
}
\arguments{
\item{n}{Le nombre de couleur souhaité. Si n > lenght palette --> n = length(palette)}

\item{show}{Logique. Si True, plot la palette.}

\item{type}{la palette de couleur choisie: "cat20", "flash12", "ws10", "gender2"}
}
\description{
Cette fonction retourne un vecteur de code de couleur en hexa. Il existe quelques palette à choix:
}
\examples{
palcol(n = 12, show = F, type = 'cat20')
palcol()
}
