% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdf.R
\name{plotdf}
\alias{plotdf}
\title{plotdf}
\usage{
plotdf(df)
}
\arguments{
\item{df}{un dataframe pour lequel un graphique est désiré}
}
\value{
Retourne un objet ggplot
}
\description{
Cette fonction prend un dataframe en argument et propose diverses options pour produire facilement un graphique avec ggplo2
}
\examples{
plotdf(df = mtcars)
plotdf()
}
