% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdf2.R
\name{plotdf2}
\alias{plotdf2}
\title{plotdf2}
\usage{
plotdf2(df)
}
\arguments{
\item{df}{un dataframe}
}
\value{
Retourne une shiny application
}
\description{
Cette fonction ouvre une application shiny et propose différents graphiques à partir d'une table SQL à choix dans l'application
}
\examples{
plotdf2(df)
plotdf2()
}
