% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdf_shine.R
\name{plotdf_shine}
\alias{plotdf_shine}
\title{plotdf_shine}
\usage{
plotdf_shine(df)
}
\arguments{
\item{df}{un data frame}
}
\value{
Retourne une shiny application
}
\description{
Cette fonction ouvre une application shiny et propose différents graphiques à partir du dataframe passé en argument
}
\examples{
plotdf_shine(df)
plotdf_shine()
}
