% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdf_shinee_sql.R
\name{plotdf_shinesql}
\alias{plotdf_shinesql}
\title{plotdf_shinesql}
\usage{
plotdf_shinesql(conn)
}
\arguments{
\item{conn}{une connexion DBI}
}
\value{
Retourne une shiny application
}
\description{
Cette fonction ouvre une application shiny et propose différents graphiques à partir d'une table SQL à choix dans l'application
}
\examples{
plotdf_shinesql(conn)
plotdf_shinesql()
}
