% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumedf.r
\name{resumedf}
\alias{resumedf}
\title{resumedf}
\usage{
resumedf(df)
}
\arguments{
\item{df}{un dataframe pour lequel un résumé est désiré}
}
\value{
Retourne un dataframe contenant le nombre de valeurs manquantes (NA, Zero, Vide), la part qu'elles représentent sur l'ensemble du dataframe.
}
\description{
This function mets a jour les données dans une table si l'id existe dans la table sql server ou alors ajoute les données a partir d'un dataframe
}
\examples{
resumedf(df = iris)
resumedf()
}
