% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumedf_sql.R
\name{resumedf_sql}
\alias{resumedf_sql}
\title{resumedf_sql}
\usage{
resumedf_sql(conn, Schema, nom.table)
}
\arguments{
\item{conn}{A DBI connection}

\item{Schema}{The sql schema}

\item{nom.table}{The sql Table}

\item{df}{un dataframe pour lequel un résumé est désiré}
}
\value{
Retourne un dataframe contenant le nombre de valeurs manquantes (NA, Zero, Vide), la part qu'elles représentent sur l'ensemble du dataframe.
}
\description{
This function mets a jour les données dans une table si l'id existe dans la table sql server ou alors ajoute les données a partir d'un dataframe
}
\examples{
resumedf_sql(conn = conn, Schema = 'dbo' , nom.table = 'mtcars')
resumedf_sql()
}
