% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine.R
\name{shinee}
\alias{shinee}
\title{plotdf}
\usage{
shinee(var)
}
\arguments{
\item{var}{une variable contine}
}
\value{
Retourne une shiny application
}
\description{
Cette fonction prend un vecteur de variable continue et retourne un histogramme de distribtion selon le nombre de classe choisie
}
\examples{
shine(df$continuesVar)
plotdf()
}
