% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_table.R
\name{stat_add}
\alias{stat_add}
\title{add_to_table}
\usage{
stat_add(dataframe = NULL, conn, BDD, Schema, Table)
}
\arguments{
\item{dataframe}{A dataframe containing the data. Default is null}

\item{conn}{A DBI connection}

\item{BDD}{The sql database name}

\item{Schema}{The sql schema}

\item{Table}{The sql Table}
}
\description{
This function add a dataframe to an existing sql table.
}
\examples{
stat_add(mtcars, conn, 'STATPRODTEMP','GEN','MTCARS')
stat_add()
}
