% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_update.R
\name{stat_update}
\alias{stat_update}
\title{stat_update}
\usage{
stat_update(df = NULL, conn, BDD, Schema, table_name, id)
}
\arguments{
\item{df}{A dataframe containing the data. Default is null}

\item{conn}{A DBI connection}

\item{BDD}{The sql database name}

\item{Schema}{The sql schema}

\item{table_name}{The sql Table}

\item{id}{a unique identifier per row.}
}
\description{
This function mets a jour les données dans une table sql server a partir d'un dataframe
}
\examples{
stat_update(df = mtcars, conn = conn, BDD = 'STATPRODTEMP',Schema = 'GEN',table_name = 'MTCARS',id = rownames(mtcars))
stat_update()
}
