% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_create.R
\name{table_create}
\alias{table_create}
\title{table_create}
\usage{
table_create(
  df = NULL,
  conn = NULL,
  BDD = NULL,
  Schema = NULL,
  table_name = NULL
)
}
\arguments{
\item{df}{A dataframe containing the data. Default is null}

\item{conn}{A DBI connection}

\item{BDD}{The sql database name}

\item{Schema}{The sql schema}

\item{table_name}{The sql Table}
}
\description{
This function crée une table a partir d'un dataframe
}
\examples{
table_create(df = mtcars, conn = conn, BDD = 'STATPRODTEMP',Schema = 'GEN',table_name = 'MTCARS')
table_create()
}
