% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_or_insert2.R
\name{update_or_insert2}
\alias{update_or_insert2}
\title{update_or_insert2}
\usage{
update_or_insert2(df, conn, BDD, Schema, table_name, id)
}
\arguments{
\item{df}{A dataframe containing the data. Default is null}

\item{conn}{A DBI connection}

\item{BDD}{The sql database name}

\item{Schema}{The sql schema}

\item{table_name}{The sql Table}

\item{id}{a unique identifier per row.}
}
\description{
This function mets a jour les données PAR BATCH de 100 dans une table si l'id existe dans la table sql server ou alors ajoute les données a partir d'un dataframe
}
\examples{
update_or_insert2(df = mtcars, conn = conn, BDD = 'STATPRODTEMP',Schema = 'GEN',table_name = 'MTCARS',id = rownames(mtcars))
update_or_insert2()
}
