% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert.R
\name{upsert}
\alias{upsert}
\title{upsert}
\usage{
upsert(df, conn, BDD, Schema, table_name, id)
}
\arguments{
\item{df}{A dataframe containing the data. Default is null}

\item{conn}{A DBI connection}

\item{BDD}{The sql database name}

\item{Schema}{The sql schema}

\item{table_name}{The sql Table}

\item{id}{a unique identifier per row.}
}
\description{
This function mets a jour les données dans une table si l'id existe dans la table sql server ou alors ajoute les données a partir d'un dataframe
}
\examples{
upsert(df = mtcars, conn = conn, BDD = 'STATPRODTEMP',Schema = 'GEN',table_name = 'MTCARS',id = rownames(mtcars))
upsert()
}
