% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_replace.r
\name{drop_replace}
\alias{drop_replace}
\title{drop_replace}
\usage{
drop_replace(
  df = NULL,
  conn = NULL,
  BDD = NULL,
  Schema = NULL,
  table_name = NULL
)
}
\arguments{
\item{df}{Un data frame contenant les nouvelles données}

\item{conn}{Une connection DBI}

\item{BDD}{Le nom de la base de donnée}

\item{Schema}{Le schéma SQL server dans lequel la table se trouve}

\item{table_name}{Le nom de la table sql}
}
\description{
cette fonction remplace les données d'une table existante
}
\examples{
drop_replace(df = mtcars, conn = conn, BDD = 'STATPRODTEMP',Schema = 'GEN',table_name = 'MTCARS')
drop_replace()
}
