% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_table.R
\name{stat_add}
\alias{stat_add}
\title{stat_add}
\usage{
stat_add(df = NULL, conn = NULL, BDD = NULL, Schema = NULL, table_name = NULL)
}
\arguments{
\item{df}{un dataframe contenant les données}

\item{conn}{une connexion DBI}

\item{BDD}{Le nom de la base de données sql}

\item{Schema}{Le schema sql}

\item{table_name}{Le nom de la table sql}
}
\description{
This function add a dataframe to an existing sql table.
}
\examples{
stat_add(mtcars, conn, 'STATPRODTEMP','GEN','MTCARS')
stat_add()
}
